+++ backend/test/Anela.Heblo.Tests/Common/HebloWebApplicationFactory.cs
@@ -10,7 +10,8 @@ using Microsoft.AspNetCore.Http;
 using Anela.Heblo.API;
 using Anela.Heblo.Persistence;
 using Anela.Heblo.API.Infrastructure.Authentication;
-using Anela.Heblo.Application.Features.KnowledgeBase.Services;
+using Anela.Heblo.Application.Shared.Rag;
+using Anela.Heblo.Application.Shared.Rag.OneDrive;
 using Anela.Heblo.Domain.Features.Catalog.Inventory;
 using Microsoft.Extensions.Hosting;
 using Hangfire;
